refactor CategoryMapper
